+++ Rain.Wave/Filters/AmplitudeWaveFilter.cs
@@ -6,16 +6,16 @@ namespace Rain.Wave.Filters
 {
 	public sealed class AmplitudeWaveFilter : IWave
 	{
-		private readonly float _maxAmplitude;
+		private readonly double _maxAmplitude;
 		private readonly IWave _baseWave;
 
-		public AmplitudeWaveFilter(float maxAmplitude, IWave baseWave)
+		public AmplitudeWaveFilter(double maxAmplitude, IWave baseWave)
 		{
 			_maxAmplitude = maxAmplitude;
 			_baseWave = baseWave;
 		}
 
-		public float Probe(float time)
+		public double Probe(double time)
 		{
 			return Math.Max(-_maxAmplitude, Math.Min(_maxAmplitude, _baseWave.Probe(time)));
 		}
