+++ Rain.Wave/Transformers/LoopWaveTransformer.cs
@@ -6,16 +6,16 @@ namespace Rain.Wave.Transformers
 {
 	public class LoopWaveTransformer : IWave
 	{
-		private readonly float _period;
+		private readonly double _period;
 		private readonly IWave _baseWave;
 
-		public LoopWaveTransformer(float period, IWave baseWave)
+		public LoopWaveTransformer(double period, IWave baseWave)
 		{
 			_period = period;
 			_baseWave = baseWave;
 		}
 
-		public float Probe(float time)
+		public double Probe(double time)
 		{
 			return _baseWave.Probe(time % _period);
 		}
